Fixed some warnings of gcc 4.9.2

+++ include/sdsl/uint256_t.hpp
@@ -47,31 +47,27 @@ class uint256_t
         inline uint256_t(uint256_t&& x):m_lo(std::move(x.m_lo)),
             m_mid(std::move(x.m_mid)), m_high(std::move(x.m_high)) {}
 
-        uint256_t& operator=(const uint256_t& x)
-        {
+        uint256_t& operator=(const uint256_t& x) {
             m_lo = x.m_lo;
             m_mid = x.m_mid;
             m_high = x.m_high;
             return *this;
         }
 
-        uint256_t& operator=(uint256_t&& x)
-        {
+        uint256_t& operator=(uint256_t&& x) {
             m_lo = std::move(x.m_lo);
             m_mid = std::move(x.m_mid);
             m_high = std::move(x.m_high);
             return *this;
         }
 
-        inline uint16_t popcount()
-        {
+        inline uint16_t popcount() {
             return ((uint16_t)bits::cnt(m_lo)) + (uint16_t) bits::cnt(m_mid)
                    + (uint16_t) bits::cnt(m_high>>64) + (uint16_t) bits::cnt(m_high);
         }
 
-        inline uint16_t hi()
-        {
-            if (m_high == 0ULL) {
+        inline uint16_t hi() {
+            if (m_high == (uint128_t)0ULL) {
                 if (m_mid) {
                     return bits::hi(m_mid) + 64;
                 } else {
@@ -87,8 +83,7 @@ class uint256_t
             }
         }
 
-        inline uint16_t select(uint32_t i)
-        {
+        inline uint16_t select(uint32_t i) {
             uint16_t x = 0;
             if ((x= (uint16_t) bits::cnt(m_lo)) >= i) {
                 return bits::sel(m_lo, i);
@@ -107,8 +102,7 @@ class uint256_t
             return bits::sel(hh, i) + 192;
         }
 
-        inline uint256_t& operator+=(const uint256_t& x)
-        {
+        inline uint256_t& operator+=(const uint256_t& x) {
             uint128_t lo = (uint128_t)m_lo + x.m_lo;
             uint128_t mid = (uint128_t)m_mid + x.m_mid + (lo >> 64);
             m_lo = lo; m_mid = mid;
@@ -117,25 +111,22 @@ class uint256_t
 //			return uint256_t(lo, mid, m_high + x.m_high + (mid >> 64));
         }
 
-        inline uint256_t operator+(const uint256_t& x)
-        {
+        inline uint256_t operator+(const uint256_t& x) {
             uint128_t lo = ((uint128_t)m_lo) + x.m_lo;
             uint128_t mid = (uint128_t)m_mid + x.m_mid + (lo >> 64);
             return uint256_t(lo, mid, m_high + x.m_high + (mid >> 64));
         }
 
-        inline uint256_t operator-(const uint256_t& x)
-        {
+        inline uint256_t operator-(const uint256_t& x) {
 //			add two's complement of x
-            uint128_t lo = (uint128_t)m_lo + (~x.m_lo) + 1ULL;
+            uint128_t lo = (uint128_t)m_lo + (~x.m_lo) + (uint128_t)1ULL;
             uint128_t mid = (uint128_t)m_mid + (~x.m_mid) + (lo >> 64);
             return uint256_t(lo, mid, m_high + (~x.m_high) + (mid >> 64));
         }
 
-        inline uint256_t& operator-=(const uint256_t& x)
-        {
+        inline uint256_t& operator-=(const uint256_t& x) {
 //			add two's complement of x
-            uint128_t lo = (uint128_t)m_lo + (~x.m_lo) + 1ULL;
+            uint128_t lo = (uint128_t)m_lo + (~x.m_lo) + (uint128_t)1ULL;
             uint128_t mid = (uint128_t)m_mid + (~x.m_mid) + (lo >> 64);
             m_lo = lo;
             m_mid = mid;
@@ -144,19 +135,16 @@ class uint256_t
         }
 
 
-        inline uint256_t operator|(const uint256_t& x)
-        {
+        inline uint256_t operator|(const uint256_t& x) {
             return uint256_t(m_lo|x.m_lo, m_mid|x.m_mid, m_high|x.m_high);
         }
 
-        inline uint256_t& operator|=(const uint256_t& x)
-        {
+        inline uint256_t& operator|=(const uint256_t& x) {
             m_lo |= x.m_lo; m_mid |= x.m_mid; m_high |= x.m_high;
             return *this;
         }
 
-        inline uint256_t operator&(const uint256_t& x)
-        {
+        inline uint256_t operator&(const uint256_t& x) {
             return uint256_t(m_lo&x.m_lo, m_mid&x.m_mid, m_high&x.m_high);
         }
         /* // is not needed since we can convert uint256_t to uint64_t
@@ -165,8 +153,7 @@ class uint256_t
                 }
         */
 
-        inline uint256_t operator<<(int x) const
-        {
+        inline uint256_t operator<<(int x) const {
             if (x < 128) {
                 uint128_t high = m_high << x;
                 uint128_t low  = (((uint128_t)m_mid<<64) | m_lo);
@@ -179,8 +166,7 @@ class uint256_t
             }
         }
 
-        inline uint256_t operator>>(int x) const
-        {
+        inline uint256_t operator>>(int x) const {
             if (x < 128) {
                 uint128_t low  = (((uint128_t)m_mid<<64) | m_lo) >> x;
                 low |= ((m_high << (127-x))<<1);
@@ -191,26 +177,22 @@ class uint256_t
             }
         }
 
-        inline uint256_t& operator=(const uint64_t& x)
-        {
+        inline uint256_t& operator=(const uint64_t& x) {
             m_high = 0;
             m_mid = 0;
             m_lo = x;
             return *this;
         }
 
-        inline bool operator==(const uint256_t& x) const
-        {
+        inline bool operator==(const uint256_t& x) const {
             return (m_lo == x.m_lo) and (m_mid == x.m_mid) and (m_high == x.m_high);
         }
 
-        inline bool operator!=(const uint256_t& x) const
-        {
+        inline bool operator!=(const uint256_t& x) const {
             return !(*this == x);
         }
 
-        inline bool operator>=(const uint256_t& x) const
-        {
+        inline bool operator>=(const uint256_t& x) const {
             if (m_high != x.m_high) {
                 return m_high > x.m_high;
             }
@@ -221,8 +203,7 @@ class uint256_t
             }
         }
 
-        inline bool operator<=(const uint256_t& x) const
-        {
+        inline bool operator<=(const uint256_t& x) const {
             if (m_high != x.m_high) {
                 return m_high < x.m_high;
             }
@@ -233,8 +214,7 @@ class uint256_t
             }
         }
 
-        inline bool operator>(const uint256_t& x) const
-        {
+        inline bool operator>(const uint256_t& x) const {
             if (m_high != x.m_high) {
                 return m_high > x.m_high;
             }
@@ -245,16 +225,14 @@ class uint256_t
             }
         }
 
-        inline bool operator>(const uint64_t& x) const
-        {
-            if (m_high > 0ULL or m_mid > 0ULL) {
+        inline bool operator>(const uint64_t& x) const {
+            if (m_high > (uint128_t)0ULL or m_mid > (uint128_t)0ULL) {
                 return true;
             }
             return m_lo > x;
         }
 
-        inline bool operator<(const uint256_t& x) const
-        {
+        inline bool operator<(const uint256_t& x) const {
             if (m_high != x.m_high) {
                 return m_high < x.m_high;
             }
@@ -265,8 +243,7 @@ class uint256_t
             }
         }
 
-        inline operator uint64_t()
-        {
+        inline operator uint64_t() {
             return m_lo;
         }
 };
